pyclient: dummy input gen in joinsplit module
